requirements.txt
-e git+https://github.com/silenius/saexts.git#egg=saexts
bcrypt==3.1.3
Beaker==1.8.1
cffi==1.10.0
Chameleon==3.2
file-magic==0.3.0
gunicorn==19.7.1
hupper==1.0
marshmallow==3.0.0b4
PasteDeploy==1.5.2
plaster==1.0
plaster-pastedeploy==0.4.1
psycopg2==2.7.1
pycparser==2.17
pyramid==1.9.1
pyramid-beaker==0.8
pyramid-chameleon==0.3
pyramid-mailer==0.15.1
pyramid-tm==2.2
pytz==2017.2
repoze.lru==0.7
repoze.sendmail==4.4.1
rutter==0.2
six==1.10.0
SQLAlchemy==1.1.15
transaction==2.1.2
translationstring==1.3
venusian==1.1.0
WebOb==1.7.4
zope.deprecation==4.3.0
zope.interface==4.4.3
zope.sqlalchemy==0.7.7
